pyproject.toml
[tool.poetry]
name = "avito"
version = "0.1.0"
description = ""
authors = ["mariazks <mariazks@yandex-team.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
psycopg2-binary = "^2.9.3"
Pillow = "^9.1.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
